use type assertions instead of reflect
You can type-assert to interfaces, even in cases where
this is "upgrading" to a more-restrictive interface which
needs additional fields, as long as the underlying concrete
type implements the interface.

I wasn't 100% sure about this, but I did some test cases
and it worked. (Then it failed because my copy of spf13/pflag
was too old to handle IP masks, but updating it fixed that.)