add support for maps
This is a bit clunky and possibly just plain a bad idea, but I
got really curious about whether it could be done. Answer: Yes.

The difficulty in supporting maps is (1) making sure we can find
names for them, (2) handling assignments to them, since their
contents aren't addressable. So we require that keys either be
string or satisfy Stringer, and we build a table of known
map-member arguments, and build a followup pass to come back and
assign those back over. This only works with Run/RunArgs.
The Flags function isn't changed, and won't support the new
feature, but FlagsMap could in theory -- it returns a cleanup
function to call to do the map assignments.

This is probably a bad idea.